# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MarkHub API
  version: 0.1.0
  description: 'Self-hosted bookmark hub — unified contract for Docker (FastAPI) and
    Cloudflare Workers. Error envelope: `{ "error": { "code", "message", "details?"
    } }`.'
servers:
- url: /
paths:
...
  /api/v1/boards:
    get:
      tags:
      - boards
      summary: List Boards
      operationId: list_boards_api_v1_boards_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    post:
      tags:
      - boards
      summary: Create Board
      operationId: create_board_api_v1_boards_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/boards/{board_id}:
    get:
      tags:
      - boards
      summary: Get Board
      operationId: get_board_api_v1_boards__board_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - boards
      summary: Patch Board
      operationId: patch_board_api_v1_boards__board_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - boards
      summary: Delete Board
      operationId: delete_board_api_v1_boards__board_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/scan:
    post:
      tags:
      - boards
      summary: Scan
      operationId: scan_api_v1_boards__board_id__scan_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/annotations:
    get:
      tags:
      - boards
      summary: Annotations
      operationId: annotations_api_v1_boards__board_id__annotations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/annotations/{aid}:
    patch:
      tags:
      - boards
      summary: Patch Ann
      operationId: patch_ann_api_v1_boards__board_id__annotations__aid__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      - name: aid
        in: path
        required: true
        schema:
          type: string
          title: Aid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/annotations/batch:
    post:
      tags:
      - boards
      summary: Batch Ann
      operationId: batch_ann_api_v1_boards__board_id__annotations_batch_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnBatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/groups:
    get:
      tags:
      - boards
      summary: Groups
      operationId: groups_api_v1_boards__board_id__groups_get
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - boards
      summary: Create Group
      operationId: create_group_api_v1_boards__board_id__groups_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/groups/reorder:
    post:
      tags:
      - boards
      summary: Reorder Groups
      operationId: reorder_groups_api_v1_boards__board_id__groups_reorder_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__api__boards__ReorderBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/export:
    post:
      tags:
      - boards
      summary: Export Board
      operationId: export_board_api_v1_boards__board_id__export_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/boards/{board_id}/import:
    post:
      tags:
      - boards
      summary: Import Board
      operationId: import_board_api_v1_boards__board_id__import_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__api__boards__ImportBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnnBatch:
      properties:
        atomic:
          type: boolean
          title: Atomic
          default: true
        items:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Items
      type: object
      required:
      - items
      title: AnnBatch
    AnnPatch:
      properties:
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        risk:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk
        price_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Price Tag
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
        secondary_group_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Secondary Group Ids
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Fields
      type: object
      title: AnnPatch
...
    BoardCreate:
      properties:
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
          default: ai_channels
        source_folder_ids:
          items:
            type: string
          type: array
          title: Source Folder Ids
          default: []
      type: object
      required:
      - name
      title: BoardCreate
    BoardPatch:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        source_folder_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Source Folder Ids
      type: object
      title: BoardPatch
...
    ExportBody:
      properties:
        format:
          type: string
          title: Format
          default: json
      type: object
      title: ExportBody
...
    GroupCreate:
      properties:
        name:
          type: string
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        keywords:
          items:
            type: string
          type: array
          title: Keywords
          default: []
        sort_order:
          type: integer
          title: Sort Order
          default: 0
      type: object
      required:
      - name
      title: GroupCreate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ScanBody:
      properties:
        mode:
          type: string
          title: Mode
          default: full
      type: object
      title: ScanBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    app__api__boards__ImportBody:
      properties:
        data:
          additionalProperties: true
          type: object
          title: Data
        merge:
          type: boolean
          title: Merge
          default: true
      type: object
      required:
      - data
      title: ImportBody
    app__api__boards__ReorderBody:
      properties:
        ordered_ids:
          items:
            type: string
          type: array
          title: Ordered Ids
      type: object
      required:
      - ordered_ids
      title: ReorderBody
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer